+++ lib/streamlit/delta_generator.py
@@ -16,14 +16,12 @@
 
 import functools
 import json
-import random
 import textwrap
 import numbers
 import re
 from datetime import datetime
 from datetime import date
 from datetime import time
-from typing import Optional, Any
 from datetime import timedelta
 from datetime import timezone
 
@@ -32,7 +30,6 @@ from streamlit import config
 from streamlit import cursor
 from streamlit import type_util
 from streamlit.report_thread import get_report_ctx
-from streamlit.errors import DuplicateWidgetID
 from streamlit.errors import StreamlitAPIException, StreamlitDeprecationWarning
 from streamlit.errors import NoSessionContext
 from streamlit.file_util import get_encoded_file_data
@@ -48,6 +45,16 @@ from streamlit.proto.Slider_pb2 import Slider
 from streamlit.proto.TextInput_pb2 import TextInput
 from streamlit.logger import get_logger
 from streamlit.type_util import is_type, ensure_iterable
+
+from streamlit.elements.utils import _get_widget_ui_value, _set_widget_id
+from streamlit.elements.balloons import BalloonsMixin
+from streamlit.elements.button import ButtonMixin
+from streamlit.elements.markdown import MarkdownMixin
+from streamlit.elements.text import TextMixin
+from streamlit.elements.alert import AlertMixin
+from streamlit.elements.json import JsonMixin
+from streamlit.elements.doc_string import HelpMixin
+from streamlit.elements.exception_proto import ExceptionMixin
 
 LOGGER = get_logger(__name__)
 
@@ -129,121 +136,6 @@ def _with_element(method):
     return wrapped_method
 
 
-def _build_duplicate_widget_message(
-    widget_func_name: str, user_key: Optional[str] = None
-) -> str:
-    if user_key is not None:
-        message = textwrap.dedent(
-            """
-            There are multiple identical `st.{widget_type}` widgets with
-            `key='{user_key}'`.
-
-            To fix this, please make sure that the `key` argument is unique for
-            each `st.{widget_type}` you create.
-            """
-        )
-    else:
-        message = textwrap.dedent(
-            """
-            There are multiple identical `st.{widget_type}` widgets with the
-            same generated key.
-
-            (When a widget is created, it's assigned an internal key based on
-            its structure. Multiple widgets with an identical structure will
-            result in the same internal key, which causes this error.)
-
-            To fix this, please pass a unique `key` argument to
-            `st.{widget_type}`.
-            """
-        )
-
-    return message.strip("\n").format(widget_type=widget_func_name, user_key=user_key)
-
-
-def _set_widget_id(
-    element_type: str,
-    element: Element,
-    user_key: Optional[str] = None,
-    widget_func_name: Optional[str] = None,
-) -> None:
-    """Set the widget id.
-
-    Parameters
-    ----------
-    element_type : str
-        The type of the widget as stored in proto.
-    element : proto
-        The proto of the element
-    user_key : str or None
-        Optional user-specified key to use for the widget ID.
-        If this is None, we'll generate an ID by hashing the element.
-    widget_func_name : str or None
-        The widget's DeltaGenerator function name, if it's different from
-        its element_type. Custom components are a special case: they all have
-        the element_type "component_instance", but are instantiated with
-        dynamically-named functions.
-
-    """
-
-    if widget_func_name is None:
-        widget_func_name = element_type
-
-    element_hash = hash(element.SerializeToString())
-    if user_key is not None:
-        widget_id = "%s-%s" % (user_key, element_hash)
-    else:
-        widget_id = "%s" % element_hash
-
-    ctx = get_report_ctx()
-    if ctx is not None:
-        added = ctx.widget_ids_this_run.add(widget_id)
-        if not added:
-            raise DuplicateWidgetID(
-                _build_duplicate_widget_message(widget_func_name, user_key)
-            )
-    el = getattr(element, element_type)
-    el.id = widget_id
-
-
-def _get_widget_ui_value(
-    element_type: str,
-    element: Element,
-    user_key: Optional[str] = None,
-    widget_func_name: Optional[str] = None,
-) -> Any:
-    """Get the widget ui_value from the report context.
-    NOTE: This function should be called after the proto has been filled.
-
-    Parameters
-    ----------
-    element_type : str
-        The type of the widget as stored in proto.
-    element : proto
-        The proto of the element
-    user_key : str
-        Optional user-specified string to use as the widget ID.
-        If this is None, we'll generate an ID by hashing the element.
-    widget_func_name : str or None
-        The widget's DeltaGenerator function name, if it's different from
-        its element_type. Custom components are a special case: they all have
-        the element_type "component_instance", but are instantiated with
-        dynamically-named functions.
-
-    Returns
-    -------
-    ui_value : any
-        The value of the widget set by the client or
-        the default value passed. If the report context
-        doesn't exist, None will be returned.
-
-    """
-    _set_widget_id(element_type, element, user_key, widget_func_name)
-    el = getattr(element, element_type)
-    ctx = get_report_ctx()
-    ui_value = ctx.widgets.get_widget_value(el.id) if ctx else None
-    return ui_value
-
-
 def _get_pandas_index_attr(data, attr):
     return getattr(data.index, attr, None)
 
@@ -305,7 +197,16 @@ for more information.
             """
 
 
-class DeltaGenerator(object):
+class DeltaGenerator(
+    AlertMixin,
+    BalloonsMixin,
+    ButtonMixin,
+    ExceptionMixin,
+    HelpMixin,
+    MarkdownMixin,
+    JsonMixin,
+    TextMixin,
+):
     """Creator of Delta protobuf messages.
 
     Parameters
@@ -351,6 +252,13 @@ class DeltaGenerator(object):
         # computed property that fetches the right cursor.
         #
         self._provided_cursor = cursor
+
+        # Change the module of all mixin'ed functions to be st.delta_generator,
+        # instead of the original module (e.g. st.elements.markdown)
+        for mixin in self.__class__.__bases__:
+            for (name, func) in mixin.__dict__.items():
+                if callable(func):
+                    func.__module__ = self.__module__
 
     def __getattr__(self, name):
         import streamlit as st
@@ -412,10 +320,11 @@ class DeltaGenerator(object):
 
         return "{}.{}.{}".format(container, path, index)
 
-    def _enqueue_new_element_delta(
+    def _enqueue(
         self,
-        marshall_element,
         delta_type,
+        element_proto,
+        return_value=None,
         last_index=None,
         element_width=None,
         element_height=None,
@@ -437,6 +346,71 @@ class DeltaGenerator(object):
             A DeltaGenerator that can be used to modify the newly-created
             element.
 
+        """
+        # Warn if we're called from within an @st.cache function
+        caching.maybe_show_cached_st_function_warning(self, delta_type)
+
+        # TODO: DELTAS_TYPES_THAT_MELT_DATAFRAMES mixins should fill last_index
+
+        # Copy the marshalled proto into the overall msg proto
+        msg = ForwardMsg_pb2.ForwardMsg()
+        msg_el_proto = getattr(msg.delta.new_element, delta_type)
+        msg_el_proto.CopyFrom(element_proto)
+
+        # Only enqueue message and fill in metadata if there's a container.
+        msg_was_enqueued = False
+        if self._container and self._cursor:
+            msg.metadata.parent_block.container = self._container
+            msg.metadata.parent_block.path[:] = self._cursor.path
+            msg.metadata.delta_id = self._cursor.index
+
+            if element_width is not None:
+                msg.metadata.element_dimension_spec.width = element_width
+            if element_height is not None:
+                msg.metadata.element_dimension_spec.height = element_height
+
+            _enqueue_message(msg)
+            msg_was_enqueued = True
+
+        if msg_was_enqueued:
+            # Get a DeltaGenerator that is locked to the current element
+            # position.
+            output_dg = DeltaGenerator(
+                container=self._container,
+                cursor=self._cursor.get_locked_cursor(
+                    delta_type=delta_type, last_index=last_index
+                ),
+            )
+        else:
+            # If the message was not enqueued, just return self since it's a
+            # no-op from the point of view of the app.
+            output_dg = self
+
+        return _value_or_dg(return_value, output_dg)
+
+    # NOTE: DEPRECATED. Will soon be replaced by _enqueue
+    def _enqueue_new_element_delta(
+        self,
+        marshall_element,
+        delta_type,
+        last_index=None,
+        element_width=None,
+        element_height=None,
+    ):
+        """Create NewElement delta, fill it, and enqueue it.
+        Parameters
+        ----------
+        marshall_element : callable
+            Function which sets the fields for a NewElement protobuf.
+        element_width : int or None
+            Desired width for the element
+        element_height : int or None
+            Desired height for the element
+        Returns
+        -------
+        DeltaGenerator
+            A DeltaGenerator that can be used to modify the newly-created
+            element.
         """
         rv = None
 
@@ -502,386 +476,6 @@ class DeltaGenerator(object):
         _enqueue_message(msg)
 
         return block_dg
-
-    @_with_element
-    def balloons(self, element):
-        """Draw celebratory balloons.
-
-        Example
-        -------
-        >>> st.balloons()
-
-        ...then watch your app and get ready for a celebration!
-
-        """
-        element.balloons.type = Balloons_pb2.Balloons.DEFAULT
-        element.balloons.execution_id = random.randrange(0xFFFFFFFF)
-
-    @_with_element
-    def text(self, element, body):
-        """Write fixed-width and preformatted text.
-
-        Parameters
-        ----------
-        body : str
-            The string to display.
-
-        Example
-        -------
-        >>> st.text('This is some text.')
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=PYxU1kee5ubuhGR11NsnT1
-           height: 50px
-
-        """
-
-        element.text.body = _clean_text(body)
-
-    @_with_element
-    def markdown(self, element, body, unsafe_allow_html=False):
-        """Display string formatted as Markdown.
-
-        Parameters
-        ----------
-        body : str
-            The string to display as Github-flavored Markdown. Syntax
-            information can be found at: https://github.github.com/gfm.
-
-            This also supports:
-
-            * Emoji shortcodes, such as `:+1:`  and `:sunglasses:`.
-              For a list of all supported codes,
-              see https://raw.githubusercontent.com/omnidan/node-emoji/master/lib/emoji.json.
-
-            * LaTeX expressions, by just wrapping them in "$" or "$$" (the "$$"
-              must be on their own lines). Supported LaTeX functions are listed
-              at https://katex.org/docs/supported.html.
-
-        unsafe_allow_html : bool
-            By default, any HTML tags found in the body will be escaped and
-            therefore treated as pure text. This behavior may be turned off by
-            setting this argument to True.
-
-            That said, we *strongly advise against it*. It is hard to write
-            secure HTML, so by using this argument you may be compromising your
-            users' security. For more information, see:
-
-            https://github.com/streamlit/streamlit/issues/152
-
-            *Also note that `unsafe_allow_html` is a temporary measure and may
-            be removed from Streamlit at any time.*
-
-            If you decide to turn on HTML anyway, we ask you to please tell us
-            your exact use case here:
-
-            https://discuss.streamlit.io/t/96
-
-            This will help us come up with safe APIs that allow you to do what
-            you want.
-
-        Example
-        -------
-        >>> st.markdown('Streamlit is **_really_ cool**.')
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=PXz9xgY8aB88eziDVEZLyS
-           height: 50px
-
-        """
-        element.markdown.body = _clean_text(body)
-        element.markdown.allow_html = unsafe_allow_html
-
-    @_with_element
-    def latex(self, element, body):
-        # This docstring needs to be "raw" because of the backslashes in the
-        # example below.
-        r"""Display mathematical expressions formatted as LaTeX.
-
-        Supported LaTeX functions are listed at
-        https://katex.org/docs/supported.html.
-
-        Parameters
-        ----------
-        body : str or SymPy expression
-            The string or SymPy expression to display as LaTeX. If str, it's
-            a good idea to use raw Python strings since LaTeX uses backslashes
-            a lot.
-
-
-        Example
-        -------
-        >>> st.latex(r'''
-        ...     a + ar + a r^2 + a r^3 + \cdots + a r^{n-1} =
-        ...     \sum_{k=0}^{n-1} ar^k =
-        ...     a \left(\frac{1-r^{n}}{1-r}\right)
-        ...     ''')
-
-        .. output::
-           https://share.streamlit.io/0.50.0-td2L/index.html?id=NJFsy6NbGTsH2RF9W6ioQ4
-           height: 75px
-
-        """
-        if type_util.is_sympy_expession(body):
-            import sympy
-
-            body = sympy.latex(body)
-
-        element.markdown.body = "$$\n%s\n$$" % _clean_text(body)
-
-    @_with_element
-    def code(self, element, body, language="python"):
-        """Display a code block with optional syntax highlighting.
-
-        (This is a convenience wrapper around `st.markdown()`)
-
-        Parameters
-        ----------
-        body : str
-            The string to display as code.
-
-        language : str
-            The language that the code is written in, for syntax highlighting.
-            If omitted, the code will be unstyled.
-
-        Example
-        -------
-        >>> code = '''def hello():
-        ...     print("Hello, Streamlit!")'''
-        >>> st.code(code, language='python')
-
-        .. output::
-           https://share.streamlit.io/0.27.0-kBtt/index.html?id=VDRnaCEZWSBCNUd5gNQZv2
-           height: 100px
-
-        """
-        markdown = "```%(language)s\n%(body)s\n```" % {
-            "language": language or "",
-            "body": body,
-        }
-        element.markdown.body = _clean_text(markdown)
-
-    @_with_element
-    def json(self, element, body):
-        """Display object or string as a pretty-printed JSON string.
-
-        Parameters
-        ----------
-        body : Object or str
-            The object to print as JSON. All referenced objects should be
-            serializable to JSON as well. If object is a string, we assume it
-            contains serialized JSON.
-
-        Example
-        -------
-        >>> st.json({
-        ...     'foo': 'bar',
-        ...     'baz': 'boz',
-        ...     'stuff': [
-        ...         'stuff 1',
-        ...         'stuff 2',
-        ...         'stuff 3',
-        ...         'stuff 5',
-        ...     ],
-        ... })
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=CTFkMQd89hw3yZbZ4AUymS
-           height: 280px
-
-        """
-        import streamlit as st
-
-        if not isinstance(body, str):
-            try:
-                body = json.dumps(body, default=lambda o: str(type(o)))
-            except TypeError as err:
-                st.warning(
-                    "Warning: this data structure was not fully serializable as "
-                    "JSON due to one or more unexpected keys.  (Error was: %s)" % err
-                )
-                body = json.dumps(body, skipkeys=True, default=lambda o: str(type(o)))
-
-        element.json.body = body
-
-    @_with_element
-    def title(self, element, body):
-        """Display text in title formatting.
-
-        Each document should have a single `st.title()`, although this is not
-        enforced.
-
-        Parameters
-        ----------
-        body : str
-            The text to display.
-
-        Example
-        -------
-        >>> st.title('This is a title')
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=SFcBGANWd8kWXF28XnaEZj
-           height: 100px
-
-        """
-        element.markdown.body = "# %s" % _clean_text(body)
-
-    @_with_element
-    def header(self, element, body):
-        """Display text in header formatting.
-
-        Parameters
-        ----------
-        body : str
-            The text to display.
-
-        Example
-        -------
-        >>> st.header('This is a header')
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=AnfQVFgSCQtGv6yMUMUYjj
-           height: 100px
-
-        """
-        element.markdown.body = "## %s" % _clean_text(body)
-
-    @_with_element
-    def subheader(self, element, body):
-        """Display text in subheader formatting.
-
-        Parameters
-        ----------
-        body : str
-            The text to display.
-
-        Example
-        -------
-        >>> st.subheader('This is a subheader')
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=LBKJTfFUwudrbWENSHV6cJ
-           height: 100px
-
-        """
-        element.markdown.body = "### %s" % _clean_text(body)
-
-    @_with_element
-    def error(self, element, body):
-        """Display error message.
-
-        Parameters
-        ----------
-        body : str
-            The error text to display.
-
-        Example
-        -------
-        >>> st.error('This is an error')
-
-        """
-        element.alert.body = _clean_text(body)
-        element.alert.format = Alert_pb2.Alert.ERROR
-
-    @_with_element
-    def warning(self, element, body):
-        """Display warning message.
-
-        Parameters
-        ----------
-        body : str
-            The warning text to display.
-
-        Example
-        -------
-        >>> st.warning('This is a warning')
-
-        """
-        element.alert.body = _clean_text(body)
-        element.alert.format = Alert_pb2.Alert.WARNING
-
-    @_with_element
-    def info(self, element, body):
-        """Display an informational message.
-
-        Parameters
-        ----------
-        body : str
-            The info text to display.
-
-        Example
-        -------
-        >>> st.info('This is a purely informational message')
-
-        """
-        element.alert.body = _clean_text(body)
-        element.alert.format = Alert_pb2.Alert.INFO
-
-    @_with_element
-    def success(self, element, body):
-        """Display a success message.
-
-        Parameters
-        ----------
-        body : str
-            The success text to display.
-
-        Example
-        -------
-        >>> st.success('This is a success message!')
-
-        """
-        element.alert.body = _clean_text(body)
-        element.alert.format = Alert_pb2.Alert.SUCCESS
-
-    @_with_element
-    def help(self, element, obj):
-        """Display object's doc string, nicely formatted.
-
-        Displays the doc string for this object.
-
-        Parameters
-        ----------
-        obj : Object
-            The object whose docstring should be displayed.
-
-        Example
-        -------
-
-        Don't remember how to initialize a dataframe? Try this:
-
-        >>> st.help(pandas.DataFrame)
-
-        Want to quickly check what datatype is output by a certain function?
-        Try:
-
-        >>> x = my_poorly_documented_function()
-        >>> st.help(x)
-
-        """
-        import streamlit.elements.doc_string as doc_string
-
-        doc_string.marshall(element, obj)
-
-    @_with_element
-    def exception(self, element, exception):
-        """Display an exception.
-
-        Parameters
-        ----------
-        exception : Exception
-            The exception to display.
-
-        Example
-        -------
-        >>> e = RuntimeError('This is an exception of type RuntimeError')
-        >>> st.exception(e)
-
-        """
-        import streamlit.elements.exception_proto as exception_proto
-
-        exception_proto.marshall(element.exception, exception)
 
     def dataframe(self, data=None, width=None, height=None):
         """Display a dataframe as an interactive table.
@@ -1801,40 +1395,6 @@ class DeltaGenerator(object):
         )
 
     @_with_element
-    def button(self, element, label, key=None):
-        """Display a button widget.
-
-        Parameters
-        ----------
-        label : str
-            A short label explaining to the user what this button is for.
-        key : str
-            An optional string to use as the unique key for the widget.
-            If this is omitted, a key will be generated for the widget
-            based on its content. Multiple widgets of the same type may
-            not share the same key.
-
-        Returns
-        -------
-        bool
-            If the button was clicked on the last run of the app.
-
-        Example
-        -------
-        >>> if st.button('Say hello'):
-        ...     st.write('Why hello there')
-        ... else:
-        ...     st.write('Goodbye')
-
-        """
-        element.button.label = label
-        element.button.default = False
-
-        ui_value = _get_widget_ui_value("button", element, user_key=key)
-        current_value = ui_value if ui_value is not None else False
-        return current_value
-
-    @_with_element
     def checkbox(self, element, label, value=False, key=None):
         """Display a checkbox widget.
 
@@ -1867,7 +1427,7 @@ class DeltaGenerator(object):
         element.checkbox.label = label
         element.checkbox.default = bool(value)
 
-        ui_value = _get_widget_ui_value("checkbox", element, user_key=key)
+        ui_value = _get_widget_ui_value("checkbox", element.checkbox, user_key=key)
         current_value = ui_value if ui_value is not None else value
         return bool(current_value)
 
@@ -1956,7 +1516,9 @@ class DeltaGenerator(object):
             str(format_func(option)) for option in options
         ]
 
-        ui_value = _get_widget_ui_value("multiselect", element, user_key=key)
+        ui_value = _get_widget_ui_value(
+            "multiselect", element.multiselect, user_key=key
+        )
         current_value = ui_value.value if ui_value is not None else default_value
         return [options[i] for i in current_value]
 
@@ -2017,7 +1579,7 @@ class DeltaGenerator(object):
         element.radio.default = index
         element.radio.options[:] = [str(format_func(option)) for option in options]
 
-        ui_value = _get_widget_ui_value("radio", element, user_key=key)
+        ui_value = _get_widget_ui_value("radio", element.radio, user_key=key)
         current_value = ui_value if ui_value is not None else index
 
         return (
@@ -2078,7 +1640,7 @@ class DeltaGenerator(object):
         element.selectbox.default = index
         element.selectbox.options[:] = [str(format_func(option)) for option in options]
 
-        ui_value = _get_widget_ui_value("selectbox", element, user_key=key)
+        ui_value = _get_widget_ui_value("selectbox", element.selectbox, user_key=key)
         current_value = ui_value if ui_value is not None else index
 
         return (
@@ -2414,7 +1976,7 @@ class DeltaGenerator(object):
         element.slider.step = step
         element.slider.data_type = data_type
 
-        ui_value = _get_widget_ui_value("slider", element, user_key=key)
+        ui_value = _get_widget_ui_value("slider", element.slider, user_key=key)
         if ui_value:
             current_value = getattr(ui_value, "value")
         else:
@@ -2507,7 +2069,7 @@ class DeltaGenerator(object):
             "server.maxUploadSize"
         )
         element.file_uploader.multiple_files = accept_multiple_files
-        _set_widget_id("file_uploader", element, user_key=key)
+        _set_widget_id("file_uploader", element.file_uploader, user_key=key)
 
         files = None
         ctx = get_report_ctx()
@@ -2583,7 +2145,9 @@ class DeltaGenerator(object):
         element.color_picker.label = label
         element.color_picker.default = str(value)
 
-        ui_value = _get_widget_ui_value("color_picker", element, user_key=key)
+        ui_value = _get_widget_ui_value(
+            "color_picker", element.color_picker, user_key=key
+        )
         current_value = ui_value if ui_value is not None else value
 
         return str(current_value)
@@ -2640,7 +2204,7 @@ class DeltaGenerator(object):
                 % type
             )
 
-        ui_value = _get_widget_ui_value("text_input", element, user_key=key)
+        ui_value = _get_widget_ui_value("text_input", element.text_input, user_key=key)
         current_value = ui_value if ui_value is not None else value
         return str(current_value)
 
@@ -2694,7 +2258,7 @@ class DeltaGenerator(object):
         if max_chars is not None:
             element.text_area.max_chars = max_chars
 
-        ui_value = _get_widget_ui_value("text_area", element, user_key=key)
+        ui_value = _get_widget_ui_value("text_area", element.text_area, user_key=key)
         current_value = ui_value if ui_value is not None else value
         return str(current_value)
 
@@ -2743,7 +2307,7 @@ class DeltaGenerator(object):
         element.time_input.label = label
         element.time_input.default = time.strftime(value, "%H:%M")
 
-        ui_value = _get_widget_ui_value("time_input", element, user_key=key)
+        ui_value = _get_widget_ui_value("time_input", element.time_input, user_key=key)
         current_value = (
             datetime.strptime(ui_value, "%H:%M").time()
             if ui_value is not None
@@ -2830,7 +2394,7 @@ class DeltaGenerator(object):
 
         element.date_input.max = date.strftime(max_value, "%Y/%m/%d")
 
-        ui_value = _get_widget_ui_value("date_input", element, user_key=key)
+        ui_value = _get_widget_ui_value("date_input", element.date_input, user_key=key)
 
         if ui_value is not None:
             value = getattr(ui_value, "data")
@@ -3028,7 +2592,9 @@ class DeltaGenerator(object):
         if format is not None:
             number_input.format = format
 
-        ui_value = _get_widget_ui_value("number_input", element, user_key=key)
+        ui_value = _get_widget_ui_value(
+            "number_input", element.number_input, user_key=key
+        )
 
         return ui_value if ui_value is not None else value
