+++ lib/tests/streamlit/delta_generator_test.py
@@ -28,7 +28,7 @@ from parameterized import parameterized
 import pandas as pd
 
 from streamlit.delta_generator import DeltaGenerator
-from streamlit.delta_generator import _build_duplicate_widget_message
+from streamlit.elements.utils import _build_duplicate_widget_message
 from streamlit.cursor import LockedCursor
 from streamlit.errors import DuplicateWidgetID
 from streamlit.errors import StreamlitAPIException
@@ -517,10 +517,10 @@ class WidgetIdText(testutil.DeltaGeneratorTestCase):
         element2 = Element()
         element2.text_input.CopyFrom(text_input2)
 
-        _set_widget_id("text_input", element1)
+        _set_widget_id("text_input", element1.text_input)
 
         with self.assertRaises(DuplicateWidgetID):
-            _set_widget_id("text_input", element2)
+            _set_widget_id("text_input", element2.text_input)
 
     def test_ids_are_diff_when_labels_are_diff(self):
         text_input1 = TextInput()
@@ -537,8 +537,8 @@ class WidgetIdText(testutil.DeltaGeneratorTestCase):
         element2 = Element()
         element2.text_input.CopyFrom(text_input2)
 
-        _set_widget_id("text_input", element1)
-        _set_widget_id("text_input", element2)
+        _set_widget_id("text_input", element1.text_input)
+        _set_widget_id("text_input", element2.text_input)
 
         self.assertNotEqual(element1.text_input.id, element2.text_input.id)
 
@@ -557,8 +557,8 @@ class WidgetIdText(testutil.DeltaGeneratorTestCase):
         element2 = Element()
         element2.text_area.CopyFrom(text_area2)
 
-        _set_widget_id("text_input", element1)
-        _set_widget_id("text_input", element2)
+        _set_widget_id("text_input", element1.text_input)
+        _set_widget_id("text_input", element2.text_input)
 
         self.assertNotEqual(element1.text_input.id, element2.text_area.id)
 
@@ -577,10 +577,10 @@ class WidgetIdText(testutil.DeltaGeneratorTestCase):
         element2 = Element()
         element2.text_input.CopyFrom(text_input2)
 
-        _set_widget_id("text_input", element1, user_key="some_key")
+        _set_widget_id("text_input", element1.text_input, user_key="some_key")
 
         with self.assertRaises(DuplicateWidgetID):
-            _set_widget_id("text_input", element2, user_key="some_key")
+            _set_widget_id("text_input", element2.text_input, user_key="some_key")
 
     def test_ids_are_diff_when_keys_are_diff(self):
         text_input1 = TextInput()
@@ -597,8 +597,8 @@ class WidgetIdText(testutil.DeltaGeneratorTestCase):
         element2 = Element()
         element2.text_input.CopyFrom(text_input2)
 
-        _set_widget_id("text_input", element1, user_key="some_key1")
-        _set_widget_id("text_input", element2, user_key="some_key2")
+        _set_widget_id("text_input", element1.text_input, user_key="some_key1")
+        _set_widget_id("text_input", element2.text_input, user_key="some_key2")
 
         self.assertNotEqual(element1.text_input.id, element2.text_input.id)
 
@@ -617,8 +617,8 @@ class WidgetIdText(testutil.DeltaGeneratorTestCase):
         element2 = Element()
         element2.text_input.CopyFrom(text_input2)
 
-        _set_widget_id("text_input", element1, user_key="some_key1")
-        _set_widget_id("text_input", element2, user_key="some_key1")
+        _set_widget_id("text_input", element1.text_input, user_key="some_key1")
+        _set_widget_id("text_input", element2.text_input, user_key="some_key1")
 
         self.assertNotEqual(element1.text_input.id, element2.text_input.id)
